# internal/delivery/http/dto/dto.go
package dto

type AuthQueryParams struct {
	GUID string `form:"guid" binding:"required"`
}

type AuthResponse struct {
	AccessToken  string `json:"access_token" binding:"required"`
	RefreshToken string `json:"refresh_token" binding:"required"`
}

type RefreshRequest struct {
	AccessToken  string `json:"access_token" binding:"required"`
	RefreshToken string `json:"refresh_token" binding:"required"`
}